+++ packages/expo/src/dom/webview-wrapper.tsx
@@ -22,8 +22,7 @@ import ExpoDomWebView from './webview/ExpoDOMWebView';
 import RNWebView from './webview/RNWebView';
 import { useDebugZeroHeight } from './webview/useDebugZeroHeight';
 
-type RawWebViewProps = React.ComponentProps<Exclude<typeof ExpoDomWebView, undefined>> &
-  React.ComponentProps<Exclude<typeof RNWebView, undefined>>;
+type RawWebViewProps = React.ComponentProps<Exclude<typeof RNWebView, undefined>>;
 
 interface Props {
   children?: any;
@@ -102,6 +101,10 @@ const RawWebView = React.forwardRef<object, Props>((props, ref) => {
     { names: [], props: {} }
   );
 
+  // Keep `initialProps` stable to prevent webview reloads when
+  // `injectedJavaScriptObject` changes.
+  const initialPropsRef = React.useRef(smartActions);
+
   // When the `marshalProps` change, emit them to the webview.
   React.useEffect(() => {
     emit({ type: '$$props', data: smartActions });
@@ -142,7 +145,7 @@ const RawWebView = React.forwardRef<object, Props>((props, ref) => {
       // Inject the top-most OS for the DOM component to read.
       EXPO_DOM_HOST_OS: process.env.EXPO_OS,
       // Inject the initial props
-      initialProps: smartActions,
+      initialProps: initialPropsRef.current,
     },
     injectedJavaScript: [
       dom?.matchContents ? getInjectBodySizeObserverScript() : null,
@@ -224,7 +227,6 @@ const RawWebView = React.forwardRef<object, Props>((props, ref) => {
           return emitError(error);
         }
       } else {
-        // @ts-expect-error: TODO(@kitten): The two types for this event will never match up, but we know they do
         dom?.onMessage?.(event);
       }
       _emitGlobalEvent({ type, data });